requirements.txt
# vLLM Server Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
vllm>=0.6.0
pyyaml>=6.0
requests>=2.31.0

# Optional: for development
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
